Load multiple binaries to flash
This commit introduces a new command-line argument --flash-binary
that allows to provide path of the file and the offset at which
it should be loaded into the flash memory. The new flag can be
used multiple times to load more than one binaries.

It is a generalization of --firmware-binary which could be used
only once and loaded the binary at the fixed offset of 0.